+++ packages/jaspr_router/lib/src/link.dart
@@ -3,8 +3,7 @@ import 'package:jaspr/jaspr.dart';
 import '../jaspr_router.dart';
 
 class Link extends StatelessComponent {
-  const Link(
-    this.children, {
+  const Link({
     required this.to,
     this.replace = false,
     this.extra,
@@ -14,10 +13,11 @@ class Link extends StatelessComponent {
     this.classes,
     this.styles,
     this.attributes,
+    this.child,
+    this.children,
     super.key,
   });
 
-  final List<Component> children;
   final String to;
   final bool replace;
   final Object? extra;
@@ -27,6 +27,8 @@ class Link extends StatelessComponent {
   final String? classes;
   final Styles? styles;
   final Map<String, String>? attributes;
+  final Component? child;
+  final List<Component>? children;
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
@@ -57,7 +59,7 @@ class Link extends StatelessComponent {
           }
         }
       },
-      children,
+      [if (child != null) child!, ...?children],
     );
   }
 }