+++ packages/jaspr_router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased minor
+
+- Changed the signature of the `Link` component, `children` is now a named parameter.
+
 ## 0.4.2
 
 - The title of the browser tab will no be properly set to the `title` of the current route.